% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{analyze_footnotes_fast}
\alias{analyze_footnotes_fast}
\title{Fast Footnote Analysis}
\usage{
analyze_footnotes_fast(footnotes, x_vars)
}
\arguments{
\item{footnotes}{Footnote specification}

\item{x_vars}{Analysis variables}
}
\value{
Footnote analysis structure
}
\description{
Efficient footnote processing with clear separation of concerns.
}
\keyword{internal}
