% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{analyze_groups_fast}
\alias{analyze_groups_fast}
\title{Fast Group Analysis}
\usage{
analyze_groups_fast(grp_var, data)
}
\arguments{
\item{grp_var}{Grouping variable name}

\item{data}{Data frame}
}
\value{
Group analysis structure
}
\description{
Streamlined group analysis with essential information only.
}
\keyword{internal}
