% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{analyze_strata_fast}
\alias{analyze_strata_fast}
\title{Fast Strata Analysis}
\usage{
analyze_strata_fast(strata, data)
}
\arguments{
\item{strata}{Strata variable name}

\item{data}{Data frame}
}
\value{
Strata analysis structure
}
\description{
Efficient stratification analysis.
}
\keyword{internal}
