% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{apply_theme_to_cell}
\alias{apply_theme_to_cell}
\title{Apply Theme to Single Cell}
\usage{
apply_theme_to_cell(cell, theme)
}
\arguments{
\item{cell}{Cell object}

\item{theme}{Theme configuration}
}
\value{
Themed cell
}
\description{
Apply Theme to Single Cell
}
