% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{apply_theme_to_cells}
\alias{apply_theme_to_cells}
\title{Apply Theme to All Cells}
\usage{
apply_theme_to_cells(blueprint, theme)
}
\arguments{
\item{blueprint}{Table1Blueprint object}

\item{theme}{Theme configuration}
}
\description{
Apply Theme to All Cells
}
