% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{as.data.frame.table1_blueprint}
\alias{as.data.frame.table1_blueprint}
\title{Convert Blueprint to Data Frame}
\usage{
\method{as.data.frame}{table1_blueprint}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A table1_blueprint object}

\item{row.names}{Row names (ignored)}

\item{optional}{Logical (ignored)}

\item{...}{Additional arguments (ignored)}
}
\value{
Data frame with evaluated cell contents
}
\description{
Converts a table1_blueprint to a data frame for further processing
}
