% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{can_use_parallel}
\alias{can_use_parallel}
\title{Check if parallel processing is available}
\usage{
can_use_parallel(num_cells, threshold = 1000)
}
\arguments{
\item{num_cells}{Number of cells to process}

\item{threshold}{Minimum cells for parallel overhead to be worthwhile}
}
\value{
Logical indicating if parallel processing should be used
}
\description{
Determines if the parallel package is available and if parallel
processing should be used.
}
\keyword{internal}
