% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_consolidated.R
\name{check_data_quality}
\alias{check_data_quality}
\title{Quick Data Quality Check}
\usage{
check_data_quality(data, variables)
}
\arguments{
\item{data}{Data frame}

\item{variables}{Variable names to check}
}
\description{
Quick Data Quality Check
}
\keyword{internal}
