% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_variables_exist}
\alias{check_variables_exist}
\title{Optimized Variable Name Checking}
\usage{
check_variables_exist(var_names, data_names)
}
\arguments{
\item{var_names}{Character vector of variable names}

\item{data_names}{Character vector of available column names}
}
\value{
Logical vector indicating which variables exist
}
\description{
Fast checking if variables exist in data frame columns
}
\keyword{internal}
