% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{convert_results_to_matrix}
\alias{convert_results_to_matrix}
\title{Convert parallel results to matrix}
\usage{
convert_results_to_matrix(results, blueprint)
}
\arguments{
\item{results}{List of cell results}

\item{blueprint}{Table1Blueprint object}
}
\value{
Matrix with evaluated cell contents
}
\description{
Takes results from parallel/serial evaluation and converts
to matrix format for rendering.
}
\keyword{internal}
