% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{create_cell_data_by_type}
\alias{create_cell_data_by_type}
\title{Create Cell Data by Type}
\usage{
create_cell_data_by_type(
  type,
  content,
  data_subset,
  computation,
  dependencies,
  format,
  cached_result,
  footnote_number,
  footnote_text
)
}
\arguments{
\item{type}{Cell type}

\item{content}{Cell content}

\item{data_subset}{Data subset expression}

\item{computation}{Computation expression}

\item{dependencies}{Dependencies vector}

\item{format}{Format list}

\item{cached_result}{Cached result}

\item{footnote_number}{Footnote number}

\item{footnote_text}{Footnote text}
}
\value{
Optimized cell data list
}
\description{
Creates optimized cell data structures tailored to each cell type.
This eliminates memory waste by only storing relevant fields.
}
\keyword{internal}
