% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{create_configured_blueprint}
\alias{create_configured_blueprint}
\title{Create and Configure Blueprint}
\usage{
create_configured_blueprint(
  dimensions,
  formula,
  options,
  data_info,
  dim_analysis
)
}
\description{
Create and Configure Blueprint
}
\keyword{internal}
