% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{create_factor_computation_totals}
\alias{create_factor_computation_totals}
\title{Create Factor Computation for Totals Column}
\usage{
create_factor_computation_totals(data)
}
\arguments{
\item{data}{Data frame (for total calculation)}
}
\value{
Computation expression
}
\description{
Creates computation expression for factor statistics in totals column.
}
\keyword{internal}
