% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{create_theme}
\alias{create_theme}
\title{Create Theme Configuration}
\usage{
create_theme(
  name,
  decimal_places = 1,
  variable_indent = 2,
  level_indent = 4,
  stratum_separator = "text",
  factor_separator = "text",
  dimension_rules = list(),
  rendering_rules = list(),
  css_properties = list()
)
}
\arguments{
\item{name}{Display name of the theme}

\item{decimal_places}{Decimal places for numeric formatting}

\item{variable_indent}{Indentation for variables}

\item{level_indent}{Indentation for factor levels}

\item{stratum_separator}{Separator type between strata}

\item{factor_separator}{Separator type between factors}

\item{dimension_rules}{List of dimension calculation rules}

\item{rendering_rules}{List of rendering functions}

\item{css_properties}{Named list of CSS properties}
}
\value{
Theme object (list with class "table1_theme")
}
\description{
Internal factory function for creating consistent theme objects.
Used to define all built-in themes.
}
\keyword{internal}
