% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_registry.R
\name{create_theme_bundle}
\alias{create_theme_bundle}
\title{Create theme bundle}
\usage{
create_theme_bundle(themes, name, description = NULL, author = NULL)
}
\arguments{
\item{themes}{List of theme objects}

\item{name}{Bundle name}

\item{description}{Bundle description}

\item{author}{Bundle author}
}
\value{
List with bundle information
}
\description{
Bundles multiple themes for distribution.
}
\details{
Themes can be bundled together and distributed as an R package
or shared directly.
}
\examples{
\dontrun{
themes <- list(
  create_custom_theme("Theme1"),
  create_custom_theme("Theme2")
)
bundle <- create_theme_bundle(themes, name = "MyThemes")
}

}
