% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{customize_theme}
\alias{customize_theme}
\title{Customize Theme}
\usage{
customize_theme(theme_name, decimal_places = NULL, css_properties = NULL)
}
\arguments{
\item{theme_name}{Name of base theme to customize}

\item{decimal_places}{Number of decimal places}

\item{css_properties}{Named list of CSS properties to override}
}
\value{
Modified theme object
}
\description{
Modifies an existing theme's properties.
}
