% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{data_quality_warning}
\alias{data_quality_warning}
\title{Data Quality Warning}
\usage{
data_quality_warning(variable_name, issue_type, details = NULL)
}
\arguments{
\item{variable_name}{Name of the problematic variable}

\item{issue_type}{Type of data quality issue}

\item{details}{Additional details}
}
\description{
Specialized warning for data quality issues
}
\keyword{internal}
