% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_variable_type}
\alias{detect_variable_type}
\title{Detect Variable Type}
\usage{
detect_variable_type(x)
}
\arguments{
\item{x}{Vector to analyze}
}
\value{
Character string describing variable type
}
\description{
Detect Variable Type
}
