% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{dim.table1_blueprint}
\alias{dim.table1_blueprint}
\title{Get Dimensions of Blueprint}
\usage{
\method{dim}{table1_blueprint}(x)
}
\arguments{
\item{x}{A table1_blueprint object}
}
\value{
Integer vector c(nrows, ncols)
}
\description{
S3 method for getting dimensions of a table1_blueprint object
}
