% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library path}

\item{pkgname}{Package name}
}
\description{
Called when the package is loaded. Initializes package environment with
theme registry and other package-level resources.
}
\keyword{internal}
