% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Package Unload Hook}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{Library path}
}
\description{
Called when the package is unloaded. Cleans up resources if needed.
}
\keyword{internal}
