% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{escape_html}
\alias{escape_html}
\title{Escape HTML Special Characters}
\usage{
escape_html(text)
}
\arguments{
\item{text}{Character string}
}
\value{
HTML-escaped string
}
\description{
Escape HTML Special Characters
}
