% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape LaTeX Special Characters}
\usage{
escape_latex(text)
}
\arguments{
\item{text}{Character string}
}
\value{
LaTeX-escaped string
}
\description{
Escape LaTeX Special Characters
}
