% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_processing.R
\name{estimate_parallel_speedup}
\alias{estimate_parallel_speedup}
\title{Estimate speedup from parallel processing}
\usage{
estimate_parallel_speedup(num_cells, num_cores = NULL)
}
\arguments{
\item{num_cells}{Number of cells to process}

\item{num_cores}{Number of cores available}
}
\value{
Estimated speedup factor (1.0 = no benefit, 2.0 = 2x faster)
}
\description{
Estimates the potential speedup for a given number of cells.
Takes into account parallel overhead and diminishing returns.
}
\keyword{internal}
