% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_cell.cell_separator}
\alias{evaluate_cell.cell_separator}
\title{Evaluate Separator Cell}
\usage{
\method{evaluate_cell}{cell_separator}(
  cell,
  data,
  env = parent.frame(),
  force_recalc = FALSE,
  blueprint = NULL
)
}
\arguments{
\item{cell}{Separator cell object}

\item{data}{Data frame (not used)}

\item{env}{Evaluation environment (not used)}

\item{force_recalc}{Force recalculation (not used)}

\item{blueprint}{Blueprint object (optional)}
}
\value{
Separator string
}
\description{
S3 method for evaluating separator cells.
}
