% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_cells_vectorized}
\alias{evaluate_cells_vectorized}
\title{Vectorized Cell Evaluation}
\usage{
evaluate_cells_vectorized(cells, data, parallel = FALSE)
}
\arguments{
\item{cells}{List of cell objects}

\item{data}{Data frame}

\item{parallel}{Logical, use parallel processing}
}
\value{
List of evaluated results
}
\description{
Vectorized Cell Evaluation
}
\keyword{internal}
