% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_footnote_cell}
\alias{evaluate_footnote_cell}
\title{Evaluate Footnote Cell}
\usage{
evaluate_footnote_cell(cell)
}
\arguments{
\item{cell}{Footnote cell}
}
\value{
Formatted footnote text
}
\description{
Evaluation for footnote cells with marker formatting.
}
\keyword{internal}
