% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_pvalue_cell}
\alias{evaluate_pvalue_cell}
\title{Evaluate P-value Cell}
\usage{
evaluate_pvalue_cell(cell, data)
}
\arguments{
\item{cell}{P-value cell}

\item{data}{Data frame}
}
\value{
P-value result
}
\description{
Evaluate P-value Cell
}
