% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_separator_cell}
\alias{evaluate_separator_cell}
\title{Evaluate Separator Cell}
\usage{
evaluate_separator_cell(cell)
}
\arguments{
\item{cell}{Separator cell}
}
\value{
Separator content
}
\description{
Simple evaluation for separator cells.
}
\keyword{internal}
