% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_static_cell}
\alias{evaluate_static_cell}
\title{Evaluate Static Cell}
\usage{
evaluate_static_cell(cell)
}
\arguments{
\item{cell}{Static cell object}
}
\value{
Character content
}
\description{
Simple evaluation for static content cells.
}
\keyword{internal}
