% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_summary_cell}
\alias{evaluate_summary_cell}
\title{Evaluate Summary Cell}
\usage{
evaluate_summary_cell(cell, data)
}
\arguments{
\item{cell}{Summary cell}

\item{data}{Data frame}
}
\value{
Summary result
}
\description{
Evaluate Summary Cell
}
