% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_registry.R
\name{export_theme_to_package}
\alias{export_theme_to_package}
\title{Export theme to R package format}
\usage{
export_theme_to_package(theme_obj, file = NULL)
}
\arguments{
\item{theme_obj}{Theme object to export}

\item{file}{Optional file path to write to}
}
\value{
Character string with R code (invisibly if file specified)
}
\description{
Creates R code that can be included in a package to register a theme.
}
\details{
The returned code can be included in an R package's R/ directory
to automatically register the theme when the package loads.
}
\examples{
\dontrun{
my_theme <- create_custom_theme("MyTheme")
code <- export_theme_to_package(my_theme)
cat(code)
}

}
