% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_paste}
\alias{fast_paste}
\title{Fast String Concatenation}
\usage{
fast_paste(..., sep = "")
}
\arguments{
\item{...}{Character vectors to concatenate}

\item{sep}{Separator string (default "")}
}
\value{
Single character string
}
\description{
Efficient string concatenation using pre-allocated character vectors
}
\keyword{internal}
