% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_unique_count}
\alias{fast_unique_count}
\title{Fast Unique Level Counting}
\usage{
fast_unique_count(x)
}
\arguments{
\item{x}{Vector to analyze}
}
\value{
Integer count of unique non-NA values
}
\description{
Efficiently counts unique levels in character/factor variables
}
\keyword{internal}
