% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{finalize_blueprint}
\alias{finalize_blueprint}
\title{Finalize Blueprint with Population and Theming}
\usage{
finalize_blueprint(blueprint, data, dimensions, theme)
}
\description{
Finalize Blueprint with Population and Theming
}
\keyword{internal}
