% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{find_unspanned_columns}
\alias{find_unspanned_columns}
\title{Find columns not covered by any spanner at any level}
\usage{
find_unspanned_columns(blueprint)
}
\arguments{
\item{blueprint}{A table1_blueprint}
}
\value{
Integer vector of column indices
}
\description{
Find columns not covered by any spanner at any level
}
\keyword{internal}
