% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{format_console_content}
\alias{format_console_content}
\title{Format Content for Console Theme}
\usage{
format_console_content(content, theme)
}
\arguments{
\item{content}{Character content}

\item{theme}{Theme configuration}
}
\value{
Formatted content
}
\description{
Format Content for Console Theme
}
