% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{format_journal_content}
\alias{format_journal_content}
\title{Format Content for Journal Themes}
\usage{
format_journal_content(content, theme)
}
\arguments{
\item{content}{Character content}

\item{theme}{Theme configuration}
}
\value{
Formatted content
}
\description{
Format Content for Journal Themes
}
