% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_number}
\alias{format_number}
\title{Format Number with Appropriate Precision (Optimized)}
\usage{
format_number(x, digits = 3)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Maximum decimal places (default 3)}
}
\value{
Formatted string
}
\description{
Format Number with Appropriate Precision (Optimized)
}
