% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_percentage}
\alias{format_percentage}
\title{Format Percentage}
\usage{
format_percentage(proportion, digits = 1)
}
\arguments{
\item{proportion}{Numeric proportion (0-1)}

\item{digits}{Number of decimal places (default 1)}
}
\value{
Formatted percentage string
}
\description{
Format Percentage
}
