% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{generate_col_names}
\alias{generate_col_names}
\title{Generate Column Names Optimized}
\usage{
generate_col_names(analyses, totals, pvalue)
}
\arguments{
\item{analyses}{Component analyses}

\item{totals}{Logical for totals column}

\item{pvalue}{Logical for p-value column}
}
\value{
Character vector of column names
}
\description{
Efficient column name generation.
}
\keyword{internal}
