% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{generate_dummy_variable_name}
\alias{generate_dummy_variable_name}
\title{Generate Dummy Variable Name}
\usage{
generate_dummy_variable_name(data)
}
\arguments{
\item{data}{Data frame to check for conflicts}
}
\value{
Unique variable name
}
\description{
Creates a unique dummy variable name that doesn't conflict with existing variables.
}
\keyword{internal}
