% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{generate_latex_column_spec}
\alias{generate_latex_column_spec}
\title{Generate LaTeX Column Specification}
\usage{
generate_latex_column_spec(ncols, theme)
}
\arguments{
\item{ncols}{Number of columns}

\item{theme}{Theme configuration}
}
\value{
Character string with column specification
}
\description{
Generate LaTeX Column Specification
}
