% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{generate_latex_theme_setup}
\alias{generate_latex_theme_setup}
\title{Generate LaTeX Theme Setup}
\usage{
generate_latex_theme_setup(theme)
}
\arguments{
\item{theme}{Theme configuration}
}
\value{
Character vector with LaTeX setup commands
}
\description{
Generate LaTeX Theme Setup
}
