% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{get_latex_rule}
\alias{get_latex_rule}
\title{Get LaTeX Rule Command}
\usage{
get_latex_rule(theme, position = "middle")
}
\arguments{
\item{theme}{Theme configuration}

\item{position}{Rule position ("top", "middle", "bottom")}
}
\value{
Character string with LaTeX rule command
}
\description{
Get LaTeX Rule Command
}
