% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{get_latex_table_environment}
\alias{get_latex_table_environment}
\title{Get LaTeX Table Environment}
\usage{
get_latex_table_environment(theme)
}
\arguments{
\item{theme}{Theme configuration}
}
\value{
Character string with table environment name
}
\description{
Get LaTeX Table Environment
}
