% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanners.R
\name{get_spanners}
\alias{get_spanners}
\title{Get the list of spanner definitions from the blueprint}
\usage{
get_spanners(blueprint)
}
\arguments{
\item{blueprint}{A table1_blueprint}
}
\value{
List of spanner definition lists
}
\description{
Get the list of spanner definitions from the blueprint
}
\keyword{internal}
