% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R, R/themes.R
\name{get_theme_decimal_places}
\alias{get_theme_decimal_places}
\title{Helper Functions}
\usage{
get_theme_decimal_places(theme_name)

get_theme_decimal_places(theme_name)
}
\arguments{
\item{theme_name}{Character theme name}
}
\value{
Number of decimal places

Integer number of decimal places
}
\description{
Get Theme Decimal Places

Retrieves the decimal place setting for a theme.
}
\keyword{internal}
