% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_registry.R
\name{get_theme_metadata}
\alias{get_theme_metadata}
\title{Get theme metadata}
\usage{
get_theme_metadata(theme_name)
}
\arguments{
\item{theme_name}{Character string with theme name}
}
\value{
List with theme metadata (name, author, version, description)
}
\description{
Retrieves metadata about a registered theme.
}
\examples{
\dontrun{
get_theme_metadata("nejm")
}

}
