% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_theme_registry}
\alias{get_theme_registry}
\title{Get Theme Registry}
\usage{
get_theme_registry()
}
\value{
Named list of all registered themes
}
\description{
Internal function to retrieve the theme registry from package environment.
}
\keyword{internal}
