% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R, R/theme_registry.R
\name{\%||\%}
\alias{\%||\%}
\title{Safe NULL-coalescing operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (used if x is NULL)}
}
\description{
Safe NULL-coalescing operator

Null-coalescing operator (used internally)
}
\keyword{internal}
