% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cached}
\alias{is_cached}
\title{Check Blueprint Cache Status}
\usage{
is_cached(blueprint, cache_key)
}
\arguments{
\item{blueprint}{Table1Blueprint object}

\item{cache_key}{Character string with cache key}
}
\value{
Logical indicating if result is cached
}
\description{
Determines if a statistical result is cached in the blueprint.
}
\keyword{internal}
