% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Check if Object is Empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if object is empty
}
\description{
Check if Object is Empty
}
\keyword{internal}
