% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{list_available_themes}
\alias{list_available_themes}
\title{List Available Themes}
\usage{
list_available_themes()
}
\value{
Character vector of theme names
}
\description{
Returns names of all available themes in the package.
}
\examples{
list_available_themes()

}
