% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_registry.R
\name{load_theme_bundle}
\alias{load_theme_bundle}
\title{Load theme bundle from file}
\usage{
load_theme_bundle(file)
}
\arguments{
\item{file}{Path to bundle file}
}
\value{
Theme bundle object
}
\description{
Loads a previously saved theme bundle.
}
\examples{
\dontrun{
bundle <- load_theme_bundle("mythemes.rds")
install_themes_from_bundle(bundle)
}

}
