% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanners.R
\name{max_spanner_level}
\alias{max_spanner_level}
\title{Compute the maximum spanner level in a blueprint}
\usage{
max_spanner_level(blueprint)
}
\arguments{
\item{blueprint}{A table1_blueprint}
}
\value{
Integer max level, or 0 if no spanners
}
\description{
Compute the maximum spanner level in a blueprint
}
\keyword{internal}
