% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{parse_and_analyze}
\alias{parse_and_analyze}
\title{Parse Formula and Analyze Dimensions (Combined)}
\usage{
parse_and_analyze(
  form,
  data,
  strata,
  missing,
  size,
  totals,
  pvalue,
  layout,
  footnotes
)
}
\description{
Parse Formula and Analyze Dimensions (Combined)
}
\keyword{internal}
