% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{populate_factor_variable}
\alias{populate_factor_variable}
\title{Populate Factor Variable Optimized}
\usage{
populate_factor_variable(
  blueprint,
  var_name,
  data,
  start_row,
  dimensions,
  theme_digits
)
}
\arguments{
\item{blueprint}{Blueprint object}

\item{var_name}{Variable name}

\item{data}{Data frame}

\item{start_row}{Starting row}

\item{dimensions}{Dimension analysis}

\item{theme_digits}{Theme decimal places}
}
\value{
Blueprint with factor cells populated
}
\description{
Efficient population of factor variable cells.
}
\keyword{internal}
