% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{populate_factor_variable_stratified}
\alias{populate_factor_variable_stratified}
\title{Populate Factor Variable in Stratified Analysis}
\usage{
populate_factor_variable_stratified(
  blueprint,
  var_name,
  stratum_data,
  start_row,
  dimensions,
  theme_config,
  current_stratum
)
}
\description{
Populate Factor Variable in Stratified Analysis
}
\keyword{internal}
