% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{populate_footnote_cells}
\alias{populate_footnote_cells}
\title{Populate Footnote Cells Optimized}
\usage{
populate_footnote_cells(blueprint, data, dimensions)
}
\arguments{
\item{blueprint}{Blueprint object}

\item{data}{Data frame}

\item{dimensions}{Dimension analysis}
}
\value{
Blueprint with footnote cells populated
}
\description{
Efficient population of footnote cells.
}
\keyword{internal}
