% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{populate_variable_cells}
\alias{populate_variable_cells}
\title{Populate Variable Cells Optimized}
\usage{
populate_variable_cells(blueprint, data, dimensions)
}
\arguments{
\item{blueprint}{Blueprint object}

\item{data}{Data frame}

\item{dimensions}{Dimension analysis}
}
\value{
Blueprint with variable cells populated
}
\description{
Efficient population of variable cells using dimension analysis results.
}
\keyword{internal}
