% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{print.cell}
\alias{print.cell}
\title{Print Method for Cell Objects}
\usage{
\method{print}{cell}(x, ...)
}
\arguments{
\item{x}{Cell object}

\item{...}{Additional arguments}
}
\value{
Invisibly returns `x`.
}
\description{
Informative display of cell objects showing type and key content.
}
