% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{print.table_dimensions}
\alias{print.table_dimensions}
\title{Print Method for Optimized Dimensions}
\usage{
\method{print}{table_dimensions}(x, ...)
}
\arguments{
\item{x}{Table dimensions object}

\item{...}{Additional arguments}
}
\value{
Invisibly returns `x`.
}
\description{
Clean display of dimension analysis results.
}
