% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_footnotes}
\alias{render_footnotes}
\title{Render Footnotes}
\usage{
render_footnotes(blueprint, theme, format)
}
\arguments{
\item{blueprint}{Table1Blueprint object}

\item{theme}{Theme configuration}

\item{format}{Output format}
}
\value{
Character vector with footnotes
}
\description{
Render Footnotes
}
