% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_html}
\alias{render_html}
\title{Render Blueprint to HTML}
\usage{
render_html(blueprint, theme = NULL)
}
\arguments{
\item{blueprint}{Table1Blueprint object}

\item{theme}{Theme configuration (optional)}
}
\value{
Character vector with HTML code
}
\description{
Render Blueprint to HTML
}
