% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_table_content_html}
\alias{render_table_content_html}
\title{Placeholder for HTML content rendering}
\usage{
render_table_content_html(blueprint, theme)
}
\description{
Placeholder for HTML content rendering
}
