% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanners.R
\name{resolve_child_columns}
\alias{resolve_child_columns}
\title{Collect all column indices owned by a set of child spanners}
\usage{
resolve_child_columns(spanners, child_ids)
}
\description{
Collect all column indices owned by a set of child spanners
}
\keyword{internal}
