% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanners.R
\name{resolve_column_indices}
\alias{resolve_column_indices}
\title{Resolve column references to integer indices}
\usage{
resolve_column_indices(columns, col_names)
}
\description{
Resolve column references to integer indices
}
\keyword{internal}
