% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zztable1.R
\name{row_pv}
\alias{row_pv}
\alias{row_pv.factor}
\alias{row_pv.numeric}
\title{Calculate p-values for comparisons}
\usage{
row_pv(x, grp, missing, ...)

\method{row_pv}{factor}(x, grp, ...)

\method{row_pv}{numeric}(x, grp, missing, ...)
}
\arguments{
\item{x}{Variable to test}

\item{grp}{Grouping variable}

\item{missing}{Logical; whether to show missing values}

\item{...}{Additional parameters}
}
\value{
Vector of p-values
}
\description{
Internal function that calculates appropriate statistical tests and p-values
based on variable type (factor, numeric, etc.)
}
\section{Methods (by class)}{
\itemize{
\item \code{row_pv(factor)}: Method for factor variables

\item \code{row_pv(numeric)}: Method for numeric variables

}}
\examples{
# For demonstration only, these functions are typically used internally
\dontrun{
  # For a factor variable:
  x <- factor(c("Yes", "No", "Yes", "No", "Yes"))
  grp <- factor(c("Treatment", "Treatment", "Placebo", "Placebo", "Placebo"))
  row_pv(x, grp)
  
  # For a numeric variable:
  y <- c(23, 45, 67, NA, 34)
  grp <- factor(c("Treatment", "Treatment", "Placebo", "Placebo", "Placebo"))
  row_pv(y, grp, missing = TRUE)
}

}
