% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe Division}
\usage{
safe_divide(numerator, denominator, na_value = NA)
}
\arguments{
\item{numerator}{Numeric numerator}

\item{denominator}{Numeric denominator}

\item{na_value}{Value to return if division by zero (default NA)}
}
\value{
Result of division or na_value
}
\description{
Safe Division
}
\keyword{internal}
