% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{[<-.table1_blueprint}
\alias{[<-.table1_blueprint}
\title{Optimized Cell Assignment for Blueprint}
\usage{
\method{[}{table1_blueprint}(x, i, j) <- value
}
\arguments{
\item{x}{A table1_blueprint object}

\item{i}{Row index (1-based)}

\item{j}{Column index (1-based)}

\item{value}{A Cell object or NULL to remove}
}
\value{
Modified table1_blueprint object
}
\description{
Provides efficient assignment of cells to blueprint positions using
hash table storage with automatic memory management.
}
\details{
Assignment automatically manages memory by:
\itemize{
  \item Storing only non-NULL cells
  \item Removing cells when assigned NULL
  \item Updating cell count metadata
  \item Validating cell objects before storage
}
}
\examples{
bp <- Table1Blueprint(5, 3)

# Assign cell
bp[1, 1] <- Cell(type = "static", content = "Variable")

# Remove cell
bp[1, 1] <- NULL

}
