% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{table1_assert}
\alias{table1_assert}
\title{Assert Function}
\usage{
table1_assert(condition, error_message, error_class = "assertion_error")
}
\arguments{
\item{condition}{Logical condition to test}

\item{error_message}{Error message if condition is FALSE}

\item{error_class}{Class for the error}
}
\description{
Assertion with standardized error reporting
}
\keyword{internal}
