% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zztable1.R
\docType{data}
\name{theme_nejm}
\alias{theme_nejm}
\title{NEJM-style theme for tables}
\format{
A list with two elements:
\describe{
\item{foreground}{A vector of foreground (text) colors}
\item{background}{A vector of background colors}
}
}
\usage{
theme_nejm
}
\description{
A predefined theme for formatting tables in the style of
the New England Journal of Medicine.
}
\examples{
# View the NEJM theme definition
theme_nejm

}
\keyword{datasets}
