% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_numeric_summary_specification}
\alias{validate_numeric_summary_specification}
\title{Validate Numeric Summary Specification}
\usage{
validate_numeric_summary_specification(numeric_summary)
}
\arguments{
\item{numeric_summary}{Summary specification}
}
\value{
Invisibly returns TRUE if valid
}
\description{
Validates numeric summary parameter which can be either a string
specifying built-in options or a custom function.
}
\keyword{internal}
