% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{validation_chain}
\alias{validation_chain}
\title{Validation Chain}
\usage{
validation_chain(value, ...)
}
\arguments{
\item{value}{Value to validate}

\item{...}{Validation functions to apply}
}
\description{
Chains multiple validation functions together
}
\keyword{internal}
