% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{apply_theme}
\alias{apply_theme}
\title{Apply Theme to Blueprint}
\usage{
apply_theme(blueprint, theme)
}
\arguments{
\item{blueprint}{Table1Blueprint object}

\item{theme}{Theme configuration list}
}
\value{
Modified blueprint with theme applied
}
\description{
Apply Theme to Blueprint
}
