% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{create_pvalue_cell}
\alias{create_pvalue_cell}
\title{Create P-value Cell Optimized}
\usage{
create_pvalue_cell(var_name, grp_var, test_type, data = NULL)
}
\arguments{
\item{var_name}{Variable name}

\item{grp_var}{Grouping variable name}

\item{test_type}{Type of statistical test}
}
\value{
P-value cell object
}
\description{
Creates optimized p-value computation cell.
}
\keyword{internal}
