% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{evaluate_computation_cell}
\alias{evaluate_computation_cell}
\title{Evaluate Computation Cell (Optimized)}
\usage{
evaluate_computation_cell(cell, data, env, force_recalc, blueprint = NULL)
}
\arguments{
\item{cell}{Computation cell object}

\item{data}{Data frame}

\item{env}{Evaluation environment}

\item{force_recalc}{Force recalculation flag}
}
\value{
Computed result
}
\description{
Streamlined evaluation with reduced overhead
}
\keyword{internal}
