% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{\%||\%}
\alias{\%||\%}
\title{Safe NULL-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (used if x is NULL)}
}
\description{
Safe NULL-coalescing operator
}
\keyword{internal}
