% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_registry.R
\name{print.theme_bundle}
\alias{print.theme_bundle}
\title{Print method for theme bundles}
\usage{
\method{print}{theme_bundle}(x, ...)
}
\arguments{
\item{x}{Theme bundle object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for theme bundles
}
