% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_table_content_console}
\alias{render_table_content_console}
\title{Placeholder for console content rendering}
\usage{
render_table_content_console(blueprint, theme)
}
\description{
Placeholder for console content rendering
}
