% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rows.R
\name{resolve_summary_columns}
\alias{resolve_summary_columns}
\title{Resolve which columns to summarise}
\usage{
resolve_summary_columns(columns, blueprint)
}
\description{
Resolve which columns to summarise
}
\keyword{internal}
