% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{summary.cell}
\alias{summary.cell}
\title{Summary Method for Cell Objects}
\usage{
\method{summary}{cell}(object, ...)
}
\arguments{
\item{object}{Cell object}

\item{...}{Additional arguments}
}
\description{
Detailed summary of cell objects including metadata.
}
