% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_formula_structure}
\alias{validate_formula_structure}
\title{Validate Formula Structure}
\usage{
validate_formula_structure(formula, data)
}
\arguments{
\item{formula}{Formula object}

\item{data}{Data frame}
}
\value{
List with validation results
}
\description{
Validate Formula Structure
}
